♻️ [Refactoring]: #455 Phase 1 guard カテゴリを tests/guard.rs に分離
先頭バイトガード / Empty / EOF / 委譲のテスト 5 件を
literal_string/tests/guard.rs に切り出し、tests.rs ハブに
`mod guard;` 宣言を追加。
